refactor GCP secret verification
